Text corrections: spelling, grammar, clarity
* Fixed: Text correction: shure -> sure

* Fixed: Text correction: past tense -> present tense

A conditional imperative on an event like this ought to be in the present tense. The past tense implies conditioning on a property, e.g., a tab having the property of having been closed (at some arbitrary undefined point in the past, not necessarily right now).

* Fixed: Text correction: parallelism

The different items on the list were written in different ways. It's mildly arguable which format to match in the case of the permissions list since each item was originally written in a different format, but it's a minor choice in the end, and the consistency improves understandability.

* Fixed: Text correction: of -> in

* Fixed: Text correction: overflowed -> overflowing

The present participle of "overflow" is "overflowing".

* Fixed: Text correction: missing word "to"

* Fixed: Text correction: addon -> add-on

It's arguable what the globally correct form is, but Firefox itself uses "add-on," so it might as well be consistent with that one.

* Fixed: Text correction: capitalization of acronyms

* Fixed: Text correction: clarity

These are mildly opinionated changes in the interest of improving the clarity of some text snippets.