Don't reopen during POSTs.
This can happen, for example, when something loads a file:// URL, that
then has static form data and javascript to immediately submit it.

We then turn the POST with data in the default container into a GET with
no data in the target container.

That doesn't work very well.

As such, let's just refuse to reopen a tab in a new container during a
POST, we can't bring the data with us (that I've figured out how to do
anyhow).