Expand secondary modifier list.
Currently, the mozilla documentation states that for secondary
modifiers:

If supplied, this must be either "Shift" or (for Firefox ≥ 63) any one
of "Ctrl", "Alt", "Command", or "MacCtrl". Must not be the modifier
already used as the main modifier.

This commit adds Ctrl, Command, and MacCtrl to the list of allowed
secondary modifiers, allowing stuff like MacCtrl+Ctrl+Period as a
binding.

Of course, that renders quite confusingly (Ctrl+Ctrl+Period), but at
least we can now accept the bindings.