Render MacCtrl as Up Arrowhead
Instead of rendering MacCtrl as Ctrl, which can cause confusion when
Command gets mapped to Ctrl instead of Command, this instead renders it
as UTF 2303 Up Arrowhead (⌃), which is the character that MacOS appears
to use for showing macctrl in bindings.

I'm not sure why they are using Up Arrowhead instead of circumflex
accent (^), but, eh, it works.